all_applications_runner/requirements.txt
streamlit>=1.28.0
requests>=2.31.0
psutil>=5.9.0
pydantic>=2.0.0
stripe>=6.0.0
python-dotenv>=1.0.0